+++ irohad/subscription/dispatcher.hpp
@@ -1,0 +1,19 @@
+/**
+ * Copyright Soramitsu Co., Ltd. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef IROHA_SUBSCRIPTION_DISPATCHER_HPP
+#define IROHA_SUBSCRIPTION_DISPATCHER_HPP
+
+#include "subscription/common.hpp"
+
+namespace iroha::subscription {
+
+  class Dispatcher final : utils::NoCopy, utils::NoMove {
+
+  };
+
+}
+
+#endif//IROHA_SUBSCRIPTION_DISPATCHER_HPP
